Test: change tests accordingly ProductValidator moved
